Modified some of the code for GetWishlistGameData so that I could also get the steam ID of the game; this is necessary for the price history lookup Added GetLowestPrices method which takes a string steamGameID: this method works but should be cleaned up because I ran into an issue where the website started refusing to return data because "too many requests". This will work in small batches (tested 6 in one go, fine... tested 191... all went to hell).
Notes:
There are several schools of thought to the above issue. I could process in batches of n; or only look up the price history for those games on sale; or separate it out entirely so that I only load the game data and let the user query for the price history on the individual result.

Also need to write the async calls for this method as it locks everything up (as expected)
And I want to also look into implementing some event changed code to update the user on the progess

Superficially, eventually want to add some graphical features to visibly show what games are on sale and have peak price points in the DataGrid...etc.